requirements.txt
azure-functions
pillow
